Rename duration summary file
